Refactor directories
This directory structure makes a bit more sense. Previously
the modules directory had Python modules pertaining to the linear
networks and the convolutional networks. This will decrease burden
on people trying to navigate the codebase.
