+++ common/darts/modules/conv/cell.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 
 from darts.api import Model
-from darts.modules.mixed_layer import MixedLayer
+from darts.modules.conv.mixed_layer import MixedLayer
 from darts.modules.operations.convolution import ConvBlock, FactorizedReduce
